==> 2.4 complexity.rmd <==
---
title: "2.4 complexity"
author: "TG"
date: "2023-03-01"
output: html_document
---


```{r}
library(tidyverse)
library(stringr)
#library(Hmisc)
library(vegan)
library(plotrix)
library(sjPlot)
library(sjmisc)
#library(gridExtra)
#library(ggiraphExtra)
library(lme4)
library(lmerTest)
library(MuMIn)
library(glmmTMB)
library(DHARMa)
library(ggeffects)
library(DarkDiv)

load("G:/My Drive/PHD/chapter 3 MPA dd/Meddata-DD/complexity_clean_data.Rdata")

plot_dir<-"G:\\My Drive\\PHD\\chapter 3 MPA dd\\Meddata-DD\\plots\\complexity model plots"

```

# final data orgenaizing

```{r}
comp<-complexity_clean_data

comp$sp.length<-NULL

comp<-comp %>%  group_by(across(1:29)) %>% summarise(amount = sum(sp.n)) %>% ungroup()

comp_wide<-spread(comp,species,amount,fill = 0)

```

# Dark diversity estimation

```{r}
sp_matrix<-comp_wide[,29:ncol(comp_wide)]
   
sp_matrix<-sp_matrix[, colSums(sp_matrix != 0) > 0] # Remove species with all 0
  
dd<-DarkDiv(sp_matrix,method = "Hypergeometric")

dd_dark<-dd[["Dark"]]
dd_dark<-as.data.frame(dd_dark)
   
dd_dark[dd_dark<0.9]<-0 # Set threshold
dd_dark[dd_dark>0]<-1 # convert to 1 sp adbove the th

dd_dark<-dd_dark %>% mutate("dark"=rowSums(dd_dark,na.rm = T))
dd_dark_sum<- dd_dark %>% select(dark)


dd_pool<-dd[["Pool"]]
dd_pool<-as.data.frame(dd_pool)
  
dd_pool[dd_pool<0.9]<-0
dd_pool[dd_pool>0]<-1
   
dd_pool<-dd_pool %>% mutate("total"=rowSums(dd_pool,na.rm = T))
dd_pool_sum<-dd_pool %>% select(total)


comp_dd<-comp_wide[,1:28]
comp_dd<-as.data.frame(comp_dd)

comp_dd<-bind_cols(comp_dd,dd_pool_sum)
comp_dd<-bind_cols(comp_dd,dd_dark_sum)

comp_dd$observed_richness<-comp_dd$total-comp_dd$dark

```

# data prepartion for model

```{r}
comp_dd$enforcement<-as.character(comp_dd$enforcement)
comp_dd$enforcement<-as.numeric(comp_dd$enforcement)

data_for_model<-comp_dd %>% filter(observed_richness >1)
data_for_model$protection<-as.factor(data_for_model$protection)

```

clean environment

```{r}
rm(list = setdiff(ls(), c("comp_dd","comp","data_for_model","plot_dir")))
completness_dir<-"G:\\My Drive\\PHD\\chapter 3 MPA dd\\Meddata-DD\\plots\\completness"

```

# Models

## Vertical relif


```{r}
completness<-data_for_model %>% filter(dark != 0) %>% 
  mutate(completness = log(observed_richness/dark))

hist(completness$completness)

completness_model<-glmmTMB(completness ~  VerRelief + protection+ (1|site), data=completness,family = gaussian())

summary(completness_model)

plot(simulateResiduals(completness_model))

plot_model(completness_model, type = "pred", terms = c("VerRelief","protection"),allow.new.levels=TRUE)

gg_comp=ggpredict(completness_model,terms = c('VerRelief[-2:4 by=0.5]'))

ggplot()+
  theme_classic()+
  geom_line(data = gg_comp,aes(x= x,y = predicted),size=1.5,color= "red")+
  geom_ribbon(data = gg_comp,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.2,fill="red")+
  xlab("Vertical relief") +
  ylab("Completeness")+
  theme(axis.text.x = element_text(size=30),
        plot.title = element_text(size=34),
        axis.text.y = element_text(size =30),
        axis.title.x = element_text(size=32),
        axis.title.y = element_text(size=32),
        legend.title = element_text(size=32),
        legend.text = element_text(size=30),
        legend.position = "none",
        strip.text.x = element_text(size = 30))

#ggsave("vr_completness.png",path = completness_dir,width = 12,height = 7)
#ggsave("vr_completness.pdf",path = completness_dir,width = 12,height = 7)

```

### Dark

```{r}
vr_dark<-glmmTMB(log(dark+1) ~  VerRelief + protection+ (1|site), data=data_for_model,family = gaussian())

summary(vr_dark)

plot(simulateResiduals(vr_dark))

plot_model(vr_dark, type = "pred", terms = c("VerRelief","protection"),allow.new.levels=TRUE)

gg_dark_vr=ggpredict(vr_dark,terms = c('VerRelief[-2:4 by=0.5]'))

ggplot()+
  theme_classic()+
  geom_line(data = gg_dark_vr,aes(x= x,y = predicted),size=1.5,color= "#7DCDE4")+
  geom_ribbon(data = gg_dark_vr,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.2,fill="#7DCDE4")+
  xlab("Vertical relief") +
  ylab("Dark diversity")+
  theme(axis.text.x = element_text(size=30),
        plot.title = element_text(size=34),
        axis.text.y = element_text(size =30),
        axis.title.x = element_text(size=32),
        axis.title.y = element_text(size=32),
        legend.title = element_text(size=32),
        legend.text = element_text(size=30),
        legend.position = "none",
        strip.text.x = element_text(size = 30))

#ggsave("vr_dark.png",path = plot_dir,width = 12,height = 7)
#ggsave("vr_dark.pdf",path = plot_dir,width = 12,height = 7)
```

### Observed


```{r}
vr_observed<-glmmTMB(log(observed_richness) ~  VerRelief + protection+ (1|site), data=data_for_model,family = gaussian())

summary(vr_observed)

plot(simulateResiduals(vr_observed))


gg_obs_vr=ggpredict(vr_observed,terms = c('VerRelief[-2:4 by=0.5]'))


ggplot()+
  theme_classic()+
  geom_line(data = gg_obs_vr,aes(x= x,y = predicted),size=1.5,color= "#F2E96C")+
  geom_ribbon(data = gg_obs_vr,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.2,fill="#F2E96C")+
  xlab("Vertical relief") +
  ylab("Observed richness")+
  theme(axis.text.x = element_text(size=30),
        plot.title = element_text(size=34),
        axis.text.y = element_text(size =30),
        axis.title.x = element_text(size=32),
        axis.title.y = element_text(size=32),
        legend.title = element_text(size=32),
        legend.text = element_text(size=30),
        legend.position = "none",
        strip.text.x = element_text(size = 30))

#ggsave("vr_observed.png",path = plot_dir,width = 12,height = 7)
#ggsave("vr_observed.pdf",path = plot_dir,width = 12,height = 7)
```

### Total

```{r}
vr_total<-glmmTMB(log(total) ~  VerRelief + protection + (1|site), data=data_for_model,family = gaussian())

summary(vr_total)

plot(simulateResiduals(vr_total))


gg_total_vr=ggpredict(vr_total,terms = c('VerRelief[-2:4 by=0.5]'))


ggplot()+
  theme_classic()+
  geom_line(data = gg_total_vr,aes(x= x,y = predicted),size=1.5,color= "#ACD672")+
  geom_ribbon(data = gg_total_vr,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.2,fill="#ACD672")+
  xlab("Vertical relief") +
  ylab("Total species pool")+
  theme(axis.text.x = element_text(size=30),
        plot.title = element_text(size=34),
        axis.text.y = element_text(size =30),
        axis.title.x = element_text(size=32),
        axis.title.y = element_text(size=32),
        legend.title = element_text(size=32),
        legend.text = element_text(size=30),
        legend.position = "none",
        strip.text.x = element_text(size = 30))

#ggsave("vr_total.png",path = plot_dir,width = 12,height = 7)
#ggsave("vr_total.pdf",path = plot_dir,width = 12,height = 7)

```


```{r}
#tab_model(vr_total,vr_observed,vr_dark,completness_model, dv.labels = c("Total", "Observed","Dark","Completness"),file = "vr_results.xls")
```



## Curvature

```{r}


completness_curve<-glmmTMB(completness ~  Curvature + protection+ (1|site), data=completness,family = gaussian())

summary(completness_curve)

plot(simulateResiduals(completness_curve))

plot_model(completness_curve, type = "pred", terms = c("Curvature","protection"),allow.new.levels=TRUE)

gg_comp_curve=ggpredict(completness_curve,terms = c('Curvature[-2:4 by=0.5]'))

ggplot()+
  theme_classic()+
  geom_line(data = gg_comp_curve,aes(x= x,y = predicted),size=1.5,color= "red")+
  geom_ribbon(data = gg_comp_curve,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.2,fill="red")+
  xlab("Curvature") +
  ylab("Completeness")+
  theme(axis.text.x = element_text(size=30),
        plot.title = element_text(size=34),
        axis.text.y = element_text(size =30),
        axis.title.x = element_text(size=32),
        axis.title.y = element_text(size=32),
        legend.title = element_text(size=32),
        legend.text = element_text(size=30),
        legend.position = "none",
        strip.text.x = element_text(size = 30))

#ggsave("Curvature_comp.png",path = completness_dir,width = 12,height = 7)
#ggsave("Curvature_comp.pdf",path = completness_dir,width = 12,height = 7)
```


### Dark

```{r}
curv_dark<-glmmTMB(log(dark+1) ~  Curvature + protection + (1|site), data=data_for_model,family = gaussian())

summary(curv_dark)

plot(simulateResiduals(curv_dark))

plot_model(curv_dark, type = "pred", terms = c("Curvature","protection"),allow.new.levels=TRUE)

gg_dark_curv=ggpredict(curv_dark,terms = c('Curvature[-2:4 by=0.5]'))

ggplot()+
  theme_classic()+
  geom_line(data = gg_dark_curv,aes(x= x,y = predicted),size=1.5,color= "#7DCDE4")+
  geom_ribbon(data = gg_dark_curv,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.2,fill="#7DCDE4")+  xlab("Curvature") +
  ylab("Dark diversity")+
  theme(axis.text.x = element_text(size=30),
        plot.title = element_text(size=34),
        axis.text.y = element_text(size =30),
        axis.title.x = element_text(size=32),
        axis.title.y = element_text(size=32),
        legend.title = element_text(size=32),
        legend.text = element_text(size=30),
        legend.position = "none",
        strip.text.x = element_text(size = 30))

#ggsave("Curvature_dark.png",path = plot_dir,width = 12,height = 7)
#ggsave("Curvature_dark.pdf",path = plot_dir,width = 12,height = 7)

```


### Observed

```{r}
curv_observed<-glmmTMB(log(observed_richness) ~  Curvature + protection + (1|site), data=data_for_model,family = gaussian())

summary(curv_observed)

plot(simulateResiduals(curv_observed))

plot_model(curv_observed, type = "pred", terms = c("Curvature","protection"),allow.new.levels=TRUE)


gg_obs_curv=ggpredict(curv_observed,terms = c('Curvature[-2:4 by=0.5]'))

ggplot()+
  theme_classic()+
  geom_line(data = gg_obs_curv,aes(x= x,y = predicted),size=1.5,color= "#F2E96C")+
  geom_ribbon(data = gg_obs_curv,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.2,fill="#F2E96C")+
  xlab("Curvature") +
  ylab("Observed richness")+
  theme(axis.text.x = element_text(size=30),
        plot.title = element_text(size=34),
        axis.text.y = element_text(size =30),
        axis.title.x = element_text(size=32),
        axis.title.y = element_text(size=32),
        legend.title = element_text(size=32),
        legend.text = element_text(size=30),
        legend.position = "none",
        strip.text.x = element_text(size = 30))

#ggsave("Curvature_obs.png",path = plot_dir,width = 12,height = 7)
#ggsave("Curvature_obs.pdf",path = plot_dir,width = 12,height = 7)
```


### Total


```{r}
curv_total<-glmmTMB(log(total) ~  Curvature + protection + (1|site), data=data_for_model,family = gaussian())

summary(curv_total)

plot(simulateResiduals(curv_total))

plot_model(curv_total, type = "pred", terms = c("Curvature","protection"),allow.new.levels=TRUE)

gg_total_curv=ggpredict(curv_total,terms = c('Curvature[-2:4 by=0.5]'))

ggplot()+
  theme_classic()+
  geom_line(data = gg_total_curv,aes(x= x,y = predicted),size=1.5,color= "#ACD672")+
  geom_ribbon(data = gg_total_curv,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.2,fill="#ACD672")+
  xlab("Curvature") +
  ylab("Total species pool")+
  theme(axis.text.x = element_text(size=30),
        plot.title = element_text(size=34),
        axis.text.y = element_text(size =30),
        axis.title.x = element_text(size=32),
        axis.title.y = element_text(size=32),
        legend.title = element_text(size=32),
        legend.text = element_text(size=30),
        legend.position = "none",
        strip.text.x = element_text(size = 30))

#ggsave("Curvature_total.png",path = plot_dir,width = 12,height = 7)
#ggsave("Curvature_total.pdf",path = plot_dir,width = 12,height = 7)
```

models summary table

```{r}
#tab_model(curv_total,curv_observed,curv_dark,completness_curve, dv.labels = c("Total", "Observed","Dark","Completness"),file = "curve_results.xls")

```
