==> 4.3 only maps.rmd <==
---
title: "4.3 only maps"
author: "TG"
date: "2023-06-13"
output: html_document
---


```{r}
library(lubridate)
library(ggplot2)
library(stringr)
library(rfishbase)
library(Hmisc)
library(vegan)
library(mobr)
library(viridis)
library(plotrix)
library(sjmisc)
library(sf)
library(ggspatial)
library(rgeos)
library(maptools)
library(dplyr)
library(tidyr)
library(leaflet)
#library(smoothr)
library(rmapshaper)

```

```{r}
select <- dplyr::select

load("G:/My Drive/PHD/chapter 3 MPA dd/Meddata-DD/clean_dd_medata.rdata")
```



## Study site map

```{r}
med_seas <- st_read("Med_World_Seas_IHO_v3_MarineRegions/Medit_seas_IHO.shp")

plot(med_seas)

# Opening the graphical device
pdf("map_test_2.pdf")

# Creating a plot
plot(med_seas$geometry)

# Closing the graphical device
dev.off() 


```


get only distinct point with different lat and lon:

```{r}
unique_coords<-medata %>% distinct(lat,lon,enforcement,protection,.keep_all = T) %>% dplyr::select(country,new_site_name,lat,lon,protection,enforcement)

```

#,position=position_jitter(h=0.12, w=0.12)
```{r}


ggplot(unique_coords) +
  geom_sf(data = med_seas$geometry, mapping = aes(),fill= "azure",color = NA) + ##D4ECFF
  geom_point(data =unique_coords,aes(x = lon, y = lat, color = enforcement), 
             size = 1.5,position=position_jitter(h=0.06, w=0.06)) + 
  xlab("Longitude") + 
  ylab("Latitude") +
   annotation_scale(location = "bl", width_hint = 0.4) +
    annotation_north_arrow(location = "bl", which_north = "true", 
        pad_x = unit(0.0, "in"), pad_y = unit(0.2, "in"))+
  theme(panel.background = element_rect(fill = "grey90"),##FFFAF7
        text=element_text(size=24, family = "serif"),
        panel.grid.major = element_line(colour = "transparent"))+
    geom_sf_text(data =med_seas, aes(label = med_seas$NAME),
                 family = "serif")+
  scale_color_manual("Protection level",values=c("#C42E1C","#D4B04A","#70A34E","#446441"),
                     labels = c("Not protected","Minimally protected","Partly protection","Fully protected"))
  
#ggsave(filename = "med_map_v1.svg", width=15, height=10)
#ggsave(filename = "med_map_v1.pdf", width=20, height=15,dpi = 1000)
#ggsave(filename = "med_map_v1.png", width=15, height=10)
```


```{r}
ggplot(unique_coords) +
  geom_sf(data = med_seas$geometry, mapping = aes(),fill= "azure",color = NA)+
  theme(axis.text.x=element_blank(),
      axis.ticks.x=element_blank(),
      axis.text.y=element_blank(),
      axis.ticks.y=element_blank())+
   theme(panel.background = element_rect(fill = "grey90"),##FFFAF7
        text=element_text(size=24, family = "serif"),
        panel.grid.major = element_line(colour = "transparent"))

#ggsave("backround_map.png",width=20, height=15,dpi = 1000)
```


```{r}

ggplot(unique_coords) +
  geom_sf(data = med_seas$geometry, mapping = aes()) + ##D4ECFF
  geom_point(data =unique_coords,aes(x = lon, y = lat, color = enforcement), 
             size = 1.5,position=position_jitter(h=0.06, w=0.06)) + 
  xlab("Longitude") + 
  ylab("Latitude") +
   annotation_scale(location = "bl", width_hint = 0.4)

ggsave(filename = "med_map_test1.pdf", width=15, height=10)

ggplot(unique_coords) +
  geom_sf(data = med_seas$geometry, mapping = aes()) + ##D4ECFF
  geom_point(data =unique_coords,aes(x = lon, y = lat, color = enforcement), 
             size = 1.5,position=position_jitter(h=0.06, w=0.06)) + 
  xlab("Longitude") + 
  ylab("Latitude") +
   annotation_scale(location = "bl", width_hint = 0.4) +
    annotation_north_arrow(location = "bl", which_north = "true", 
        pad_x = unit(0.0, "in"), pad_y = unit(0.2, "in"))+
  theme(panel.background = element_rect(fill = "grey90"),##FFFAF7
        text=element_text(size=24, family = "serif"),
        panel.grid.major = element_line(colour = "transparent"))+
    geom_sf_text(data =med_seas, aes(label = med_seas$NAME),
                 family = "serif")+
  scale_color_manual("Protection level",values=c("#C42E1C","#D4B04A","#70A34E","#446441"),
                     labels = c("Not protected","Minimally protected","Partly protection","Fully protected"))
  
```






Banyuls

```{r}
banyus_cropped <- st_crop(med_seas, xmin = 3.1, xmax = 3.18,
                                    ymin = 42.46, ymax = 42.53)


banyus<-unique_coords %>% filter(country=="France")

ggplot() +
  geom_sf(data = banyus_cropped,fill= "azure")+ ##D4ECFF
  geom_point(data =banyus,aes(x = lon, y = lat, color = enforcement), 
             size = 2)+
  theme_bw()+ggtitle("Banyuls MPA, France")+
  theme(panel.background = element_rect(fill = "grey90",color = "black"),##FFFAF7
        text=element_text(size=20,family = "serif"),legend.position = "none",
        panel.grid.major = element_line(colour = "transparent"))+
        scale_color_manual("Protection level",values=c("#C42E1C","#446441"))+
  annotation_scale(location = "bl", width_hint = 0.4)+
      annotation_north_arrow(location = "bl", which_north = "true", 
        pad_x = unit(0.0, "in"), pad_y = unit(0.2, "in"))+
    scale_x_continuous(expand = c(0, 0))+
    scale_y_continuous(expand = c(0, 0))

#ggsave(filename = "Banyuls.png", width=10, height=10)
#ggsave(filename = "Banyuls.pdf", width=10, height=10)
```




Medes

```{r}

Medes<-unique_coords %>% filter(new_site_name=="medes")

Medes_cropped <- st_crop(med_seas, xmin = 3.169, xmax = 3.26,
                                    ymin = 42.03, ymax = 42.11)




ggplot() +
  geom_sf(data = Medes_cropped,fill= "azure")+ ##D4ECFF
  geom_point(data =Medes,aes(x = lon, y = lat, color = enforcement), 
             size = 2)+
  theme_bw()+ggtitle("Medes MPA, Spain")+
  theme(panel.background = element_rect(fill = "grey90",color = "black"),##FFFAF7
        text=element_text(size=20,family = "serif"),legend.position = "none",
        panel.grid.major = element_line(colour = "transparent"))+
        scale_color_manual("Protection level",values=c("#C42E1C","#446441"))+
  annotation_scale(location = "bl", width_hint = 0.4)+
      annotation_north_arrow(location = "bl", which_north = "true", 
        pad_x = unit(0.0, "in"), pad_y = unit(0.2, "in"))+
    scale_x_continuous(expand = c(0, 0))+
    scale_y_continuous(expand = c(0, 0))

#ggsave(filename = "Medes.png", width=10, height=10)
#ggsave(filename = "Medes.pdf", width=10, height=10)
```



Brijoni

```{r}

brijuni<-unique_coords %>% filter(new_site_name=="brijuni")

brijuni_cropped <- st_crop(med_seas, xmin = 13.7, xmax = 13.85,
                                    ymin = 44.83, ymax = 44.95)

#brijuni_cropped_smooth <-ms_simplify(brijuni_cropped, keep =1)
#brijuni_cropped_smooth <- smoother::smoth(brijuni_cropped,method = "chaikin") 

ggplot() +
  geom_sf(data = brijuni_cropped,fill= "azure")+ ##D4ECFF
  geom_point(data =brijuni,aes(x = lon, y = lat, color = enforcement), 
             size = 2)+
  theme_bw()+ggtitle("Brijuni MPA, Croatia")+
  theme(panel.background = element_rect(fill = "grey90",color = "black"),##FFFAF7
        text=element_text(size=20,family = "serif"),legend.position = "none",
        panel.grid.major = element_line(colour = "transparent"))+
        scale_color_manual("Protection level",values=c("#C42E1C","#446441"))+
  annotation_scale(location = "bl", width_hint = 0.4)+
      annotation_north_arrow(location = "bl", which_north = "true", 
        pad_x = unit(0.0, "in"), pad_y = unit(0.2, "in"))+
    scale_x_continuous(expand = c(0, 0))+
    scale_y_continuous(expand = c(0, 0))

#ggsave(filename = "brijuni.png", width=10, height=10)
#ggsave(filename = "brijuni.pdf", width=10, height=10)
```




```{r}


a<-ggplot(unique_coords) +
  geom_sf(data = med_seas$geometry, mapping = aes(),fill= "azure",color = NA) + 
  geom_point(data =unique_coords,aes(x = lon, y = lat, color = enforcement,size = 2),position=position_jitter(h=0.06, w=0.06)) + 
  annotation_scale(location = "bl", width_hint = 0.4) +
  annotation_north_arrow(location = "tl", which_north = "true", 
        pad_x = unit(0.0, "in"), pad_y = unit(0.2, "in"))+
  theme(legend.position = "none",
        panel.grid.major = element_line(colour = "transparent"))+
        scale_color_manual("Protection level",values=c("#C42E1C","#D4B04A","#70A34E","#446441"))

a


b_bounds<-st_bbox(banyus_cropped$geometry)
c_bounds<-st_bbox(Medes_cropped$geometry)
d_bounds<-st_bbox(brijuni_cropped$geometry)

b_central_lat <- (b_bounds[[4]] + b_bounds[[2]]) / 2
b_central_lon <- (b_bounds[[3]] + b_bounds[[1]]) / 2

c_central_lat <- (c_bounds[[4]] + c_bounds[[2]]) / 2
c_central_lon <- (c_bounds[[3]] + c_bounds[[1]]) / 2

d_central_lat <- (d_bounds[[4]] + d_bounds[[2]]) / 2
d_central_lon <- (d_bounds[[3]] + d_bounds[[1]]) / 2


b<-ggplot() +
  geom_sf(data = banyus_cropped,fill= "azure")+ 
  geom_point(data =banyus,aes(x = lon, y = lat, color = enforcement),size = 3)+
  theme_bw()+
  theme(panel.background = element_rect(fill = "grey90",color = "black"),
       legend.position = "none",axis.title =  element_blank(),
        panel.grid.major = element_line(colour = "transparent"))+
        scale_color_manual("Protection level",values=c("#C42E1C","#446441"))+
  annotation_scale(location = "bl", width_hint = 0.4)+
    scale_x_continuous(expand = c(0, 0), breaks = round(b_central_lon,2)) +
  scale_y_continuous(expand = c(0, 0), breaks = round(b_central_lat,2))
    
b


c <-ggplot() +
  geom_sf(data = Medes_cropped,fill= "azure")+ 
  geom_point(data =Medes,aes(x = lon, y = lat, color = enforcement),size = 3)+
  theme_bw()+
  theme(panel.background = element_rect(fill = "grey90",color = "black"),
        legend.position = "none",axis.title =  element_blank(),
        panel.grid.major = element_line(colour = "transparent"))+
        scale_color_manual("Protection level",values=c("#C42E1C","#446441"))+
  annotation_scale(location = "bl", width_hint = 0.4)+
      scale_x_continuous(expand = c(0, 0), breaks = round(c_central_lon,2)) +
  scale_y_continuous(expand = c(0, 0), breaks = round(c_central_lat,2))

c


d<-ggplot() +
  geom_sf(data = brijuni_cropped,fill= "azure")+ ##D4ECFF
  geom_point(data =brijuni,aes(x = lon, y = lat, color = enforcement),size =3)+
  theme_bw()+
  theme(panel.background = element_rect(fill = "grey90",color = "black"),##FFFAF7
       legend.position = "none",
       axis.title =  element_blank(),
        panel.grid.major = element_line(colour = "transparent"))+
        scale_color_manual("Protection level",values=c("#C42E1C","#446441"))+
  annotation_scale(location = "bl", width_hint = 0.4)+
      scale_x_continuous(expand = c(0, 0), breaks = round(d_central_lon,2)) +
  scale_y_continuous(expand = c(0, 0), breaks = round(d_central_lat,2))
d


```


`


```{r}
library(patchwork)

combined_plot <- a / (b + c + d) +
  plot_layout(ncol = 1,nrow = 3,widths = c(14,7,7,7))
   
# Print the combined plot


ggsave(combined_plot,filename = "combined_map.pdf", width=10, height=10)
```
