==> 7.2 darkdiv tests no rare.rmd <==
---
title: "7.2 darkdiv tests no rare"
author: "TG"
date: "2023-02-06"
output:
   prettydoc::html_pretty:
    theme: cayman
    highlight: github
---


```{r}
library(DarkDiv)
library(tidyverse)
library(vegan)
library(plotrix)

select <- dplyr::select

```

```{r}
load("G:/My Drive/PHD/chapter 3 MPA dd/Meddata-DD/clean_dd_medata.rdata")

clean_dd_medata<-medata # keep the original

medata<-medata %>% select(1:22) # keep only metadate and abundances

medata<-medata %>% group_by(across(1:21)) %>% summarise(amount = sum(sp.n)) %>% ungroup() # sum ind from the same sp in each transect

medata$alt_basin<- ifelse(medata$country == "Israel","Israel",medata$alt_basin)

medata$enforcement<-as.character(medata$enforcement)

medata$enforcement<- ifelse(medata$site == "Kavo Gkreko" & medata$enforcement == "3" ,"1",medata$enforcement)

medata$enforcement<- ifelse(medata$site == "Akrotiri" & medata$enforcement == "3" ,"1",medata$enforcement)


#medata<- medata %>% group_by(alt_basin,species) %>% mutate(occupancy = n())  

#medata<- medata %>% filter(occupancy >0)

#medata$occupancy<-NULL

#medata_wide<-spread(medata,species,amount,fill = 0) # convert to wide format

```



```{r}
index_list<-list()
data_list<-list()

for (i in unique(medata$alt_basin)) {
 
   basin_data<-medata %>% filter(alt_basin == i)
   
    rare_sp<- basin_data %>% group_by(species) %>% mutate(occupancy = n())  

    rare_sp<- rare_sp %>% filter(occupancy < 7) %>% select(species,occupancy) %>% distinct(species)

   basin_wide<-spread(basin_data,species,amount,fill = 0)
   
   sp_matrix<-basin_wide[,21:ncol(basin_wide)]
   
   sp_matrix<-sp_matrix[, colSums(sp_matrix != 0) > 0] # Remove species with all 0
   
   rare_sp<-rare_sp %>% filter(species %in% colnames(sp_matrix))
  
   dd<-DarkDiv(sp_matrix,method = "Hypergeometric")
   
   #dd<-DarkDiv(sp_matrix,method = "ThresholdBeals",limit = "outlier",const = 0.05)
   #dd<-DarkDiv(sp_matrix,method = "Favorability",wa =F)
   
   index_list[[i]]<-dd
   
   dd_dark<-dd[["Dark"]]
   dd_dark<-as.data.frame(dd_dark)
   dd_dark<-dd_dark %>% select(-rare_sp$species)
  
  # dd_dark[dd_dark<0.9]<-0
  # dd_dark[dd_dark>0]<-1
  
   dd_dark<-dd_dark %>% mutate("dark"=rowSums(dd_dark,na.rm = T))
   dd_dark_sum<- dd_dark %>% select(dark)
   
   dd_pool<-dd[["Pool"]]
   dd_pool<-as.data.frame(dd_pool)
   dd_pool<-dd_pool %>% select(-rare_sp$species)
   
   #dd_pool[dd_pool<0.9]<-0
  # dd_pool[dd_pool>0]<-1
   
   dd_pool<-dd_pool %>% mutate("total"=rowSums(dd_pool,na.rm = T))
   dd_pool_sum<-dd_pool %>% select(total)
   
   medata_dd<-basin_wide[,1:20]
   medata_dd<-as.data.frame(medata_dd)

   medata_dd<-bind_cols(medata_dd,dd_pool_sum)
   medata_dd<-bind_cols(medata_dd,dd_dark_sum)

   medata_dd$observed_richness<-medata_dd$total-medata_dd$dark
   
   data_list[[i]]<-medata_dd
}

dd_data<-bind_rows(data_list)
```


plot by enforcement


plot by enforcement and basin

```{r}
enf_basin <- gather(dd_data,type,richness,21:23)


enf_basin <- enf_basin %>%
  group_by(enforcement,alt_basin,type) %>%
  summarise("mean_richness"= mean(richness),
            "median"=median(richness),
            "se"=std.error(richness),
            "n"=n()) 

enf_basin %>%   filter(type != "total") %>% 
ggplot()+
       aes(x=enforcement,y = mean_richness,fill = type )+
  geom_bar(stat="identity")+ylab("Mean richness per transect")+
  scale_fill_manual(values = c("#1B2D48","#F7D54A"))+theme_classic()+ facet_wrap(~alt_basin)+
  geom_text(aes(label = n, y = max(mean_richness)+1))



enf_basin %>%   filter(type == "dark") %>% 
ggplot()+
       aes(x=enforcement,y = mean_richness,fill = type )+
  geom_bar(stat="identity",position=position_dodge(.9),fill = "#1B2D48")+
  geom_errorbar(aes(ymin=mean_richness-se, ymax=mean_richness+se), width=.2,
                 position=position_dodge(.9))+ylab("Mean Dark diversity per transect") +theme_classic()+ facet_wrap(~alt_basin)+
  geom_text(aes(label = n, y = max(mean_richness)+1))


```


by site:


```{r}
by_site <- gather(dd_data,type,richness,21:23)

by_site <- by_site %>%
  group_by(enforcement,alt_basin,country,site,type) %>%
  summarise("site_richness"= mean(richness),
         "transect_n"=n()) %>%
  ungroup()

by_site_basin<-by_site %>%
  group_by(alt_basin,enforcement,type) %>%
  mutate("basin_enf_mean" = mean(site_richness),"site_num_in_basin" = n()) %>%
  ungroup()

by_site_all<-by_site %>%
  group_by(enforcement,type) %>%
  summarise("total_mean" = mean(site_richness),
            "se" = std.error(site_richness),
            "sites_enf"= n_distinct(site)) %>%
  ungroup()

#plot for the basins

by_site_basin %>%  filter(type == "dark") %>% 
ggplot()+
       aes(x=enforcement,y = basin_enf_mean, fill = type )+
  geom_bar(stat="identity",position=position_dodge(.9),fill = "#1B2D48")+
  theme_classic()+facet_wrap(~alt_basin)+
   geom_text(aes(label = site_num_in_basin, y = max(basin_enf_mean)+1))



by_site_all %>%  filter(type == "dark") %>% 
ggplot()+
       aes(x=enforcement,y = total_mean, fill = type )+
  geom_bar(stat="identity",position=position_dodge(.9),fill = "#1B2D48")+
  theme_classic()+
   geom_text(aes(label = sites_enf, y = max(total_mean)+1))


by_site_all %>% filter(type != "total") %>% 
ggplot()+
       aes(x=enforcement,y = total_mean,fill = type )+
  geom_bar(stat="identity")+ylab("Mean richness per transect")+
  scale_fill_manual(values = c("#1B2D48","#F7D54A"))+theme_classic()+ 
  geom_text(aes(label = sites_enf, y =30))


```
