==> 7.4 glmm darkdiv.rmd <==
---
title: "7.4 glmm darkdiv"
author: "TG"
date: "2023-02-14"
output: html_document
---

```{r}
library(DarkDiv)
library(tidyverse)
library(vegan)
library(plotrix)
library(sjPlot)
library(sjmisc)
library(gridExtra)
library(ggiraphExtra)
library(lme4)
library(lmerTest)
library(MuMIn)
library(glmmTMB)
library(DHARMa)
library(ggeffects)

select <- dplyr::select

#load("G:/My Drive/PHD/chapter 3 MPA dd/Meddata-DD/clean_dd_medata.rdata")

```

```{r}

clean_dd_medata<-medata # keep the original

medata<-medata %>% select(1:23) # keep only metadate and abundances

medata<-medata %>% group_by(across(1:22)) %>% summarise(amount = sum(sp.n)) %>% ungroup() # sum ind from the same sp in each transect

medata$alt_basin<- ifelse(medata$country == "Israel","Israel",medata$alt_basin) # Sep Israel

medata<-medata %>% filter(country != "Cyprus") # Remove cyprus data

medata$alt_basin<-ifelse(medata$alt_basin == "Levantine Sea","Mediterranean Sea - Eastern Basin",medata$alt_basin)# restore turkey transect to previos basin

medata_wide<-spread(medata,species,amount,fill = 0) # convert to wide format


medata_wide$alt_basin[medata_wide$new_site_name=="cabrera"]<-"Balearic (Iberian Sea)" # one basin for the site

medata_wide<-medata_wide %>% filter(new_site_name !="Crete_Kalydon")

```



```{r}
index_list<-list()
data_list<-list()

for (i in unique(medata_wide$alt_basin)) {
 
   basin_data<-medata_wide %>% filter(alt_basin == i)
   
   sp_matrix<-basin_data[,22:ncol(basin_data)]
   
   sp_matrix<-sp_matrix[, colSums(sp_matrix != 0) > 0] # Remove species with all 0
  
   dd<-DarkDiv(sp_matrix,method = "Hypergeometric")
   
   index_list[[i]]<-dd
   
   dd_dark<-dd[["Dark"]]
   dd_dark<-as.data.frame(dd_dark)
   
   dd_dark[dd_dark<0.9]<-0 # Set threshold
   dd_dark[dd_dark>0]<-1 # convert to 1 sp adbove the th
   
   dd_dark<-dd_dark %>% mutate("dark"=rowSums(dd_dark,na.rm = T))
   dd_dark_sum<- dd_dark %>% select(dark)
   
   dd_pool<-dd[["Pool"]]
   dd_pool<-as.data.frame(dd_pool)
  
   dd_pool[dd_pool<0.9]<-0
   dd_pool[dd_pool>0]<-1
   
   dd_pool<-dd_pool %>% mutate("total"=rowSums(dd_pool,na.rm = T))
   dd_pool_sum<-dd_pool %>% select(total)
   
   medata_dd<-basin_data[,1:21]
   medata_dd<-as.data.frame(medata_dd)

   medata_dd<-bind_cols(medata_dd,dd_pool_sum)
   medata_dd<-bind_cols(medata_dd,dd_dark_sum)

   medata_dd$observed_richness<-medata_dd$total-medata_dd$dark
   
   data_list[[i]]<-medata_dd
}

dd_data<-bind_rows(data_list)
```


##models at the transect level:

data organizing


```{r}

# enforcement as numeric

dd_data$enforcement<-as.character(dd_data$enforcement)
dd_data$enforcement<-as.numeric(dd_data$enforcement)

dd_data$age.reserve.yr<-as.character(dd_data$age.reserve.yr)
dd_data$age.reserve.yr<-as.numeric(dd_data$age.reserve.yr)

# convert vars to factors

dd_data$new_site_name<-as.factor(dd_data$new_site_name)
dd_data$alt_basin<-as.factor(dd_data$alt_basin)
dd_data$country<-as.factor(dd_data$country)


# remove transect with less than 3 species in the observed 

data_for_model<-dd_data %>% filter(observed_richness >1)

hist(data_for_model$observed_richness)
hist(data_for_model$dark)

# remove 0 enforcement

dd_data_no_0<-data_for_model %>% filter(enforcement != 0)

dd_data_no_0$new_site_name<-droplevels(dd_data_no_0$new_site_name)


hist(dd_data_no_0$observed_richness)
hist(dd_data_no_0$dark)



ggplot(data = data_for_model,aes(x=as.factor(enforcement),y=dark))+ geom_violin() 
ggplot(data = data_for_model,aes(x=as.factor(enforcement),y=dark))+ geom_boxplot()
ggplot(data = data_for_model,aes(x=as.factor(enforcement),y=dark))+ geom_boxplot()+facet_wrap(.~alt_basin)
ggplot(data = dd_data_no_0,aes(x=as.factor(enforcement),y=dark))+ geom_violin() 
ggplot(data = dd_data_no_0,aes(x=as.factor(enforcement),y=dark))+ geom_boxplot() 

dd_data_no_0 %>% group_by(enforcement) %>% summarise(mean= mean(dark),var = var(dark))

t<-data_for_model %>% group_by(enforcement,alt_basin) %>% summarise(mean= mean(dark),var = var(dark),n=n())

ggplot(data = t,aes(x=n,y = mean,color = as.factor(enforcement)))+geom_point()+geom_smooth(method = "lm")

data_for_model %>% group_by(enforcement,alt_basin) %>% summarise(mean= mean(observed_richness),var = var(observed_richness),n=n())


transect_per_altbasin_protection <- medata %>% group_by(alt_basin) %>% summarise("number_of_trans"= n_distinct(unique_trans_id)) 
```

```{r}
dd_data_no_0$enforcement_factor<- as.factor(dd_data_no_0$enforcement)

dd_data_no_0$enforcement_factor<-ordered(dd_data_no_0$enforcement_factor, levels = c("1", "2", "3"))

str(dd_data_no_0$enforcement_factor)
is.ordered(dd_data_no_0$enforcement_factor)

```


```{r}

nested <- glmmTMB(dark ~ enforcement + (1|alt_basin:new_site_name),
                  data = dd_data_no_0,family = nbinom1())

summary(nested)
diagnose(nested)

plot(simulateResiduals(nested))


plot_model(nested, type = "pred", terms = c("enforcement"),allow.new.levels=TRUE)


gg_nested=ggpredict(nested,terms = c('enforcement[1:3 by=1]'))


ggplot()+
  theme_classic()+
  geom_line(data = gg_nested,aes(x= x,y = predicted),size=1.5,color= "darkblue")+
  geom_ribbon(data = gg_nested,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.1,fill="darkblue")+
  xlab("Enforcement") + ylab("Dark diversity")


res_nested = simulateResiduals(nested)

plotResiduals(res_nested, form = dd_data_no_0$enforcement,asFactor = F)
plot(res_nested)
testOutliers(res_nested, type = c("bootstrap"))

plot_model(nested,type="diag")


env <- glmmTMB(dark ~ enforcement +
                 scale(tmean, center = TRUE) +
                 scale(pp_mean,center = TRUE) +
                 scale(sal_mean,center = T),
                data = dd_data_no_0,family = nbinom2)

summary(env)
diagnose(env)


plot_model(env, type = "pred", terms = c("enforcement","tmean"),allow.new.levels=TRUE)
plot_model(env, type = "pred", terms = c("enforcement","lon"),allow.new.levels=TRUE)
```

observed model

```{r}
nested_obs <- glmmTMB(observed_richness ~ enforcement + (1|alt_basin:new_site_name),
                  data = dd_data_no_0,family = nbinom1)

summary(nested_obs)
plot_model(nested_obs, type = "pred", terms = c("enforcement"))


gg_nested_obs=ggpredict(nested_obs,terms = c('enforcement[1:3 by=1]'))


ggplot()+
  theme_classic()+
  geom_line(data = gg_nested_obs,aes(x= x,y = predicted),size=1.5,color= "gold")+
  geom_ribbon(data = gg_nested_obs,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.1,fill="gold")+
  xlab("Enforcement") + ylab("Observed richness")


res_nested_obs = simulateResiduals(nested_obs)

testOutliers(res_nested_obs, type = c("binomial"))

plot(res_nested_obs)

plot_model(res_nested,type="diag")



env_obs <- glmmTMB(observed_richness ~ enforcement + tmean + pp_mean + sal_mean +depth,
                  data = dd_data_no_0,family = nbinom1)

summary(env_obs)
plot_model(env_obs, type = "pred", terms = c("enforcement","tmean"),allow.new.levels=TRUE)
```



```{r}
sub <- dd_data_no_0

sub$site_enf<- paste(sub$new_site_name,sub$enforcement,sep = "_")

sub <- sub %>% group_by(site_enf) %>% slice_sample(n=8)


```

```{r}
nested_sub <- glmmTMB(dark ~ enforcement + (1|alt_basin:new_site_name),
                  data = sub,family = nbinom1)

summary(nested_sub)

diagnose(nested_sub)

nested_sub_res = simulateResiduals(nested_sub)
plotSimulatedResiduals(nested_sub_res)
plot_model(nested_sub,type="diag")

plot_model(nested_sub, type = "pred", terms = c("enforcement"))


env_sub <- glmmTMB(dark ~ enforcement  +
                  scale(tmean, center = T) +
                  scale(pp_mean,center = T) +
                  scale(sal_mean,center = T),
                  data = sub,family = nbinom2)


env_res = simulateResiduals(env_sub)
plot(env_res)

summary(env_sub)
plot_model(env_sub, type = "pred", terms = c("enforcement"))

diagnose(env_sub)


```
