==> 7.5 darkdiv final.rmd <==
---
title: "7.5 darkdiv final"
author: "TG"
date: "2023-03-01"
output:
   prettydoc::html_pretty:
    theme: cayman
    highlight: github
---

```{r}
library(DarkDiv)
library(tidyverse)
library(vegan)
library(plotrix)
library(sjPlot)
library(sjmisc)
library(gridExtra)
library(lme4)
library(lmerTest)
library(MuMIn)
library(glmmTMB)
library(DHARMa)
library(ggeffects)

select <- dplyr::select

load("G:/My Drive/PHD/chapter 3 MPA dd/Meddata-DD/clean_dd_medata.rdata")

plot_dir<-"G:\\My Drive\\PHD\\chapter 3 MPA dd\\Meddata-DD\\plots\\models plots medata"

```


## Final data prepartion

```{r}

clean_dd_medata<-medata # keep the original

medata<-medata %>% select(1:23) # keep only metadate and abundances

medata<-medata %>% group_by(across(1:22)) %>% summarise(amount = sum(sp.n)) %>% ungroup() # sum ind from the same sp in each transect

medata$alt_basin<- ifelse(medata$country == "Israel","Israel",medata$alt_basin) # Sep Israel

medata<-medata %>% filter(country != "Cyprus") # Remove cyprus data

medata$alt_basin<-ifelse(medata$alt_basin == "Levantine Sea","Mediterranean Sea - Eastern Basin",medata$alt_basin)# restore turkey transect to previos basin

medata_wide<-spread(medata,species,amount,fill = 0) # convert to wide format


medata_wide$alt_basin[medata_wide$new_site_name=="cabrera"]<-"Balearic (Iberian Sea)" # one basin for the site



medata_wide<-medata_wide %>% filter(new_site_name !="Crete_Kalydon")

```


## calculate the dark diversity


```{r}
index_list<-list()
data_list<-list()

for (i in unique(medata_wide$alt_basin)) {
 
   basin_data<-medata_wide %>% filter(alt_basin == i)
   
   sp_matrix<-basin_data[,22:ncol(basin_data)]
   
   sp_matrix<-sp_matrix[, colSums(sp_matrix != 0) > 0] # Remove species with all 0
  
   dd<-DarkDiv(sp_matrix,method = "Hypergeometric")
   
   index_list[[i]]<-dd
   
   dd_dark<-dd[["Dark"]]
   dd_dark<-as.data.frame(dd_dark)
   
   dd_dark[dd_dark<0.9]<-0 # Set threshold
   dd_dark[dd_dark>0]<-1 # convert to 1 sp adbove the th
   
   dd_dark<-dd_dark %>% mutate("dark"=rowSums(dd_dark,na.rm = T))
   dd_dark_sum<- dd_dark %>% select(dark)
   
   dd_pool<-dd[["Pool"]]
   dd_pool<-as.data.frame(dd_pool)
  
   dd_pool[dd_pool<0.9]<-0
   dd_pool[dd_pool>0]<-1
   
   dd_pool<-dd_pool %>% mutate("total"=rowSums(dd_pool,na.rm = T))
   dd_pool_sum<-dd_pool %>% select(total)
   
   medata_dd<-basin_data[,1:21]
   medata_dd<-as.data.frame(medata_dd)

   medata_dd<-bind_cols(medata_dd,dd_pool_sum)
   medata_dd<-bind_cols(medata_dd,dd_dark_sum)

   medata_dd$observed_richness<-medata_dd$total-medata_dd$dark
   
   data_list[[i]]<-medata_dd
}

dd_data<-bind_rows(data_list)
```


## after darkdiv data prepartion


```{r}

# enforcement as numeric

dd_data$enforcement<-as.character(dd_data$enforcement)
dd_data$enforcement<-as.numeric(dd_data$enforcement)

dd_data$age.reserve.yr<-as.character(dd_data$age.reserve.yr)
dd_data$age.reserve.yr<-as.numeric(dd_data$age.reserve.yr)

# convert vars to factors

dd_data$new_site_name<-as.factor(dd_data$new_site_name)
dd_data$alt_basin<-as.factor(dd_data$alt_basin)
dd_data$country<-as.factor(dd_data$country)


# remove transect with less than 2 species in the observed 

data_for_model<-dd_data %>% filter(observed_richness >1)

# remove 0 enforcement

dd_data_no_0<-data_for_model %>% filter(enforcement != 0)

dd_data_no_0$new_site_name<-droplevels(dd_data_no_0$new_site_name)

```


add enforcement as orderd factor

```{r}
dd_data_no_0$enforcement_factor<- as.factor(dd_data_no_0$enforcement)

dd_data_no_0$enforcement_factor<-ordered(dd_data_no_0$enforcement_factor, levels = c("1", "2", "3"))

str(dd_data_no_0$enforcement_factor)
is.ordered(dd_data_no_0$enforcement_factor)

```

clean environment:


```{r}

rm(list = setdiff(ls(), c("dd_data","dd_data_no_0","medata","data_for_model","plot_dir")))

```

# Models

## Dark


```{r}
Dark_model <- glmmTMB(log(dark+1) ~ enforcement + (1|alt_basin:new_site_name),
                  data = dd_data_no_0,family = gaussian())


summary(Dark_model)

plot(simulateResiduals(Dark_model))


gg_dark=ggpredict(Dark_model,terms = c('enforcement[1:3 by=1]'))

ggplot()+
  theme_classic()+
  geom_line(data = gg_dark,aes(x= x,y = predicted),size=1.5,color= "#7DCDE4")+
  geom_ribbon(data = gg_dark,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.2,fill="#7DCDE4")+
  xlab("Protection") + scale_x_continuous(limits=c(1,3),breaks = c(1,2,3)) +
  ggtitle("Dark diversity")+
  ylab("Richness")+
  theme(axis.text.x = element_text(size=30),
        plot.title = element_text(size=32),
        axis.text.y = element_text(size =28),
        axis.title.x = element_text(size=30),
        axis.title.y = element_text(size=30))


ggsave("medata_dark.png",path = plot_dir,width = 12,height = 7)
ggsave("medata_dark.pdf",path = plot_dir,width = 12,height = 7)

```


# Observed

```{r}
Observed <- glmmTMB(log(observed_richness) ~ enforcement + (1|alt_basin:new_site_name),
                  data = dd_data_no_0,family = gaussian())

summary(Observed)

plot(simulateResiduals(Observed))

gg_observed=ggpredict(Observed,terms = c('enforcement[1:3 by=1]'))


ggplot()+
  theme_classic()+
  geom_line(data = gg_observed,aes(x= x,y = predicted),size=1.5,color= "#F2E96C")+
  geom_ribbon(data = gg_observed,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.2,fill="#F2E96C")+
  xlab("Protection") + scale_x_continuous(limits=c(1,3),breaks = c(1,2,3)) +
  ggtitle("Observed richness")+
  ylab("Richness")+
  theme(axis.text.x = element_text(size=30),
        plot.title = element_text(size=32),
        axis.text.y = element_text(size =28),
        axis.title.x = element_text(size=30),
        axis.title.y = element_text(size=30))

ggsave("medata_observed.png",path = plot_dir,width = 12,height = 7)
ggsave("medata_observed.pdf",path = plot_dir,width = 12,height = 7)
```

## Total species pool



```{r}
Total <- glmmTMB(log(total) ~ enforcement + (1|alt_basin:new_site_name),
                  data = dd_data_no_0,family = gaussian())

summary(Total)

plot(simulateResiduals(Total))

gg_total=ggpredict(Total,terms = c('enforcement[1:3 by=1]'))


ggplot()+
  theme_classic()+
  geom_line(data = gg_total,aes(x= x,y = predicted),size=1.5,color= "#ACD672")+
  geom_ribbon(data = gg_total,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.2,fill="#ACD672")+
  xlab("Protection") + scale_x_continuous(limits=c(1,3),breaks = c(1,2,3)) +
  ggtitle("Total species pool")+
  ylab("Richness")+
  theme(axis.text.x = element_text(size=30),
        plot.title = element_text(size=32),
        axis.text.y = element_text(size =28),
        axis.title.x = element_text(size=30),
        axis.title.y = element_text(size=30))

ggsave("medata_total.png",path = plot_dir,width = 12,height = 7)
ggsave("medata_total.pdf",path = plot_dir,width = 12,height = 7)
```

models summry tables:

```{r}
#tab_model(Total,Observed,Dark_model, dv.labels = c("Total", "Observed","Dark"))
```


```{r}
#tab_model(Total,Observed,Dark_model, dv.labels = c("Total", "Observed","Dark"),file = "medata_results.xls")
```


# completness

```{r}

completness_dir<-"G:\\My Drive\\PHD\\chapter 3 MPA dd\\Meddata-DD\\plots\\completness"


completness<-dd_data_no_0 %>%
  mutate(completness = dark/observed_richness)

hist(completness$completness)

ggplot(completness, aes(x=completness)) +
  geom_histogram()+
    facet_grid(enforcement ~ .)


completness_model <- glmmTMB(completness ~ enforcement + (1|alt_basin:new_site_name),
                  data = completness,family = gaussian())


summary(completness_model)

plot(simulateResiduals(completness_model))


gg_comp=ggpredict(completness_model,terms = c('enforcement[1:3 by=1]'))

ggplot()+
  theme_classic()+
  geom_line(data = gg_comp,aes(x= x,y = predicted),size=1.5,color= "red")+
  geom_ribbon(data = gg_comp,aes(x=x,ymin=conf.low,ymax=conf.high),alpha=0.2,fill="red")+
  xlab("Protection") + scale_x_continuous(limits=c(1,3),breaks = c(1,2,3)) +
  ggtitle("Completness")+
  ylab("Completness")+
  theme(axis.text.x = element_text(size=30),
        plot.title = element_text(size=32),
        axis.text.y = element_text(size =28),
        axis.title.x = element_text(size=30),
        axis.title.y = element_text(size=30))



#ggsave("completness_protection.png",path = completness_dir,width = 12,height = 7)
#ggsave("completness_protection.pdf",path = completness_dir,width = 12,height = 7)


#tab_model(completness_model,file = "completness_protection.xls")

```
