==> 10.1 summary tables.rmd <==
---
title: "10.1 summary tables"
author: "TG"
date: "2023-06-18"
output: html_document
---

NOTE!!! the new table is in 9.3 script (01/11/23)

```{r}
library(DarkDiv)
library(tidyverse)
library(vegan)
library(plotrix)
library(sjPlot)
library(sjmisc)
library(gridExtra)
library(lme4)
library(lmerTest)
library(MuMIn)
library(glmmTMB)
library(DHARMa)
library(ggeffects)

select <- dplyr::select

load("G:/My Drive/PHD/chapter 3 MPA dd/Meddata-DD/clean_dd_medata.rdata")

plot_dir<-"G:\\My Drive\\PHD\\chapter 3 MPA dd\\Meddata-DD\\plots\\tables"

```


## Final data prepartion

```{r}

clean_dd_medata<-medata # keep the original

medata<-medata %>% select(1:23) # keep only metadate and abundances

medata<-medata %>% group_by(across(1:22)) %>% summarise(amount = sum(sp.n)) %>% ungroup() # sum ind from the same sp in each transect

medata$alt_basin<- ifelse(medata$country == "Israel","Israel",medata$alt_basin) # Sep Israel

medata<-medata %>% filter(country != "Cyprus") # Remove cyprus data

medata$alt_basin<-ifelse(medata$alt_basin == "Levantine Sea","Mediterranean Sea - Eastern Basin",medata$alt_basin)# restore turkey transect to previos basin

medata_wide<-spread(medata,species,amount,fill = 0) # convert to wide format


medata_wide$alt_basin[medata_wide$new_site_name=="cabrera"]<-"Balearic (Iberian Sea)" # one basin for the site



medata_wide<-medata_wide %>% filter(new_site_name !="Crete_Kalydon")

```

```{r}
medata_wide$new_site_name <- sub("_", " ", medata_wide$new_site_name)
medata_wide$new_site_name<-str_to_title(medata_wide$new_site_name)
```


# site number inside and outside MPA

```{r}

sites<-medata_wide %>% select(new_site_name,protection)%>% distinct()

transect_protection<-medata_wide %>% group_by(protection) %>% summarise(transect_num = n_distinct(unique_trans_id))

transects_enf<-medata_wide %>% group_by(enforcement) %>% summarise(transect_num = n_distinct(unique_trans_id))


transect_site<-medata_wide %>% group_by(country,new_site_name,protection) %>% summarise(transect_num = n_distinct(unique_trans_id)) %>% arrange(country,new_site_name,protection)

colnames(transect_site)<- c("Country","Site","Protection","Number of transects")

#write.csv(transect_site,paste0(plot_dir,"\\transect_per_site.csv"))
```

transect per basin


```{r}

basin<-medata_wide %>% group_by(alt_basin) %>% summarise("Number of sites"=n_distinct(new_site_name,protection), "Number of transects" = n_distinct(unique_trans_id)) %>% rename(Basin = alt_basin) %>% arrange(Basin)

#write.csv(basin,paste0(plot_dir,"\\transect_per_basin.csv"))

```
